move manifests pkg
