+++ pkg/manifests/fsscanner/manifests.go
