+++ pkg/manifests/parser/error_types.go
