+++ pkg/netpol/connlist/internal/ingressanalyzer/ingress_analyzer_test.go
@@ -9,10 +9,10 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/np-guard/netpol-analyzer/pkg/logger"
+	"github.com/np-guard/netpol-analyzer/pkg/manifests/fsscanner"
+	"github.com/np-guard/netpol-analyzer/pkg/manifests/parser"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/eval"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/internal/testutils"
-	"github.com/np-guard/netpol-analyzer/pkg/netpol/manifests/fsscanner"
-	"github.com/np-guard/netpol-analyzer/pkg/netpol/manifests/parser"
 )
 
 // helping func - scans the directory objects and returns the ingress analyzer built from them
