+++ pkg/netpol/diff/diff.go
@@ -14,11 +14,11 @@ import (
 	"k8s.io/cli-runtime/pkg/resource"
 
 	"github.com/np-guard/netpol-analyzer/pkg/logger"
+	"github.com/np-guard/netpol-analyzer/pkg/manifests/fsscanner"
+	"github.com/np-guard/netpol-analyzer/pkg/manifests/parser"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/common"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/connlist"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/eval"
-	"github.com/np-guard/netpol-analyzer/pkg/netpol/manifests/fsscanner"
-	"github.com/np-guard/netpol-analyzer/pkg/netpol/manifests/parser"
 
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 )
